requirements.txt
groq>=0.12.0
PyAudio>=0.2.14
webrtcvad-wheels>=2.0.14
pynput>=1.7.6
keyboard>=0.13.5
pystray>=0.19.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pyperclip>=1.8.2
comtypes>=1.2.0
httpx>=0.27.0
packaging>=23.0
pymorphy3>=2.0
pymorphy3-dicts-uk>=2.4
pywebview>=5.0
